Properly rebuild resource file if shader changes (#53)
Unfortunately CMAKE_AUTORCC does not properly add dependencies on the
files included by the qrc file. This means that when a file changes that
is included in the qrc file, the qrc file is not automatically rebuilt.
This results in broken builds.

Fix this by building the qrc file directly (instead of relying on
CMAKE_AUTORCC) and adding proper dependencies.

As a side cleanup, move the resources.qrc file to resources/, so the
top-level directory is cleaner.